update kernel
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@54507 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/universal/linux-4.4/fs/btrfs/ioctl.c
@@ -1960,7 +1960,7 @@ static noinline int copy_to_sk(struct btrfs_root *root,
 			       struct btrfs_path *path,
 			       struct btrfs_key *key,
 			       struct btrfs_ioctl_search_key *sk,
-			       size_t *buf_size,
+			       u64 *buf_size,
 			       char __user *ubuf,
 			       unsigned long *sk_offset,
 			       int *num_found)
@@ -2092,7 +2092,7 @@ out:
 
 static noinline int search_ioctl(struct inode *inode,
 				 struct btrfs_ioctl_search_key *sk,
-				 size_t *buf_size,
+				 u64 *buf_size,
 				 char __user *ubuf)
 {
 	struct btrfs_root *root;
@@ -2166,7 +2166,7 @@ static noinline int btrfs_ioctl_tree_search(struct file *file,
 	struct btrfs_ioctl_search_key sk;
 	struct inode *inode;
 	int ret;
-	size_t buf_size;
+	u64 buf_size;
 
 	if (!capable(CAP_SYS_ADMIN))
 		return -EPERM;
@@ -2200,8 +2200,8 @@ static noinline int btrfs_ioctl_tree_search_v2(struct file *file,
 	struct btrfs_ioctl_search_args_v2 args;
 	struct inode *inode;
 	int ret;
-	size_t buf_size;
-	const size_t buf_limit = 16 * 1024 * 1024;
+	u64 buf_size;
+	const u64 buf_limit = 16 * 1024 * 1024;
 
 	if (!capable(CAP_SYS_ADMIN))
 		return -EPERM;
